Move MeteorEvents to IPC directory.
