+++ npm-packages/meteor-vite/src/bin/worker/IPC/MeteorEvents.ts
@@ -1,6 +1,5 @@
 import { EventEmitter } from 'events';
-import pc from 'picocolors';
-import Logger from '../Logger';
+import Logger from '../../../Logger';
 
 type MeteorIPCTopic = 'webapp-reload-client' | 'webapp-pause-client' | 'client-refresh';
